# Replace stubs in ndimage by Jitify