--- a/cupyx/scipy/ndimage/_filters_core.py
+++ b/cupyx/scipy/ndimage/_filters_core.py
@@ -143,58 +143,7 @@ def _call_kernel(kernel, input, weights, output, structure=None,
     return output
 
 
-math_constants_preamble = "#include <math_constants.h>\n"
-
-
-_CAST_FUNCTION = """
-// Implements a casting function to make it compatible with scipy
-// Use like cast<to_type>(value)
-// It's actually really simple - most of this is <type_traits>
-
-// Small bit of <type_traits> which cannot be imported in NVRTC
-// Requires compiling with --std=c++11 or higher
-template<bool B, class T=void> struct enable_if {};
-template<class T> struct enable_if<true, T> { typedef T type; };
-template<class T> struct remove_const          { typedef T type; };
-template<class T> struct remove_const<const T> { typedef T type; };
-template<class T> struct remove_volatile             { typedef T type; };
-template<class T> struct remove_volatile<volatile T> { typedef T type; };
-template<class T> struct remove_cv {
-  typedef typename remove_volatile<typename remove_const<T>::type>::type type;
-};
-template<class T, T v>
-struct integral_constant { static constexpr T value = v; };
-typedef integral_constant<bool, true> true_type;
-typedef integral_constant<bool, false> false_type;
-template<class T> struct __is_fp : public false_type {};
-template<>        struct __is_fp<float16> : public true_type {};
-template<>        struct __is_fp<float> : public true_type {};
-template<>        struct __is_fp<double> : public true_type {};
-template<>        struct __is_fp<long double> : public true_type {};
-template<class T> struct is_floating_point
-    : public __is_fp<typename remove_cv<T>::type> {};
-template<class T> struct is_signed
-    : public integral_constant<bool, (T)(-1)<0> {};
-template<> struct is_signed<float16> : public true_type {};
-template<class T> struct is_signed<complex<T>> : public is_signed<T> {};
-
-template <class B, class A>
-__device__
-typename enable_if<!is_floating_point<A>::value||is_signed<B>::value, B>::type
-cast(A a) { return (B)a; }
-
-template <class B, class A>
-__device__
-typename enable_if<is_floating_point<A>::value&&!is_signed<B>::value, B>::type
-cast(A a) { return (a >= 0) ? (B)a : -(B)(-a); }
-
-template <class T>
-__device__ bool nonzero(T x) { return (bool)x; }
-
-template <typename T>
-__device__ bool nonzero(complex<T> x) { return x.real() || x.imag(); }
-
-"""
+math_constants_preamble = '#include <math_constants.h>\n'
 
 
 def _generate_nd_kernel(name, pre, found, post, mode, w_shape, int_type,
@@ -305,7 +254,8 @@ def _generate_nd_kernel(name, pre, found, post, mode, w_shape, int_type,
         name += '_with_structure'
     if has_mask:
         name += '_with_mask'
-    preamble = math_constants_preamble + _CAST_FUNCTION + preamble
+    preamble = math_constants_preamble + '#include <type_traits>\n' + preamble
+    options += ('--std=c++11', '-DCUPY_USE_JITIFY')
     return cupy.ElementwiseKernel(in_params, out_params, operation, name,
                                   reduce_dims=False, preamble=preamble,
-                                  options=('--std=c++11',) + options)
+                                  options=options)