Follow-up of #4228. We no longer need customized C++ std code now that we have Jitify.

TODO:
- [x] Test HIP