@leofang 2020-11-11
Looks like ROCm/HIP can handle `type_traits` as expected, but there're some issues that many tests fail (#4268), which is irrelevant of the present PR.