@leofang 2020-11-12
> Converted to draft and check if I can make some improvements to upstream (Jitify): [NVIDIA/jitify#82](https://github.com/NVIDIA/jitify/pull/82)

Blocked by that and also #4277. After that the JIT overhead will be very small.