@leofang 2020-11-29
I did quite a bit tests in order to evaluate the impact of Jitify overhead. It turns out that we need #4277 + the patch applied to `cupy/cuda/jitify,pyx` (included in this PR) in order to minimize the compile time, see below:

time (roughly) | before | after (no change to `jitify.pyx`) | after (with `jitify.pyx` fixed)
--------|-------|------|-------
`ndimage_tests/test_filters.py` | 10 min | 20 min | 16 min 
`ndimage_tests/` | 30 min | 45 min | 37.5 min

Note that `test_filters.py` was picked because it's affected the most, as can be seen from the increased timing. I think this PR made the overhead acceptable (there are ~5.7k tests in `test_filters.py`, so the averaged overhead, assuming each test compiles 1 new kernel, is (6*60)/5700 = 63ms per test).