@leofang 2020-12-10
Is the CI timeout shortened? In `cupy-py3` all ndimage tests actually passed, but the CI was terminated. Let me retry...

Jenkins, test this please